requirements.txt
flask==3.1.0
flask-cors==5.0.0
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2
python-dotenv==1.0.1
openai==1.64.0
langchain==0.3.18
langchain-openai==0.3.6
langchain-community==0.3.17
chromadb==0.6.3
sentence-transformers==3.4.1
pypdf==5.3.0
markdown==3.7